Refactoring: move bucketIndexReader to dedicated source file
